超越判定を外に出し、Command::Parserを使った形に書き換え
